fix(leagueMember): 리그 랭킹 백분위 Float 타입(n.m% 형식)으로 변경

+++ src/main/java/com/moayong/api/domain/league/service/LeagueService.java
@@ -46,7 +46,7 @@ public class LeagueService {
         return leagueRepository.findAll();
     }
 
-    public Integer getNextLevel(League league, int rate) {
+    public Integer getNextLevel(League league, float rate) {
         int level = league.getLevel();
 
         if (rate <= league.getPromotionRate()) return level + 1;
